requirements.txt
# Django core dependencies
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
uwsgi>=2.0.20,<2.1

# Modal and protobuf - compatible versions to resolve conflicts
protobuf>=4.21.6,<5.0.0
modal>=0.63.0
grpcio>=1.50.0,<1.63.0
grpcio-tools>=1.50.0,<1.63.0

# Review analysis dependencies
django-filter>=22.1
django-cors-headers>=3.13.0
redis>=4.3.0
celery>=5.2.0

# Additional utilities
python-dotenv>=0.20.0
requests>=2.28.0

# PhoneBridge dependencies (NEW)
cryptography>=3.4.8
pyjwt>=2.4.0

# NLP and ML dependencies for local fallback
# torch>=1.12.0
# transformers>=4.20.0
# numpy>=1.21.0
# scikit-learn>=1.1.0

# Development and debugging
python-decouple>=3.6